pyproject.toml
[project]
name = "pdf-ocr"
version = "0.2.0"
description = "Simple python script to perform OCR (Optical Character Recognition) on PDF files locally."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.0",
    "pdfplumber>=0.10.0",
    "Pillow>=11.0.0",
]
